+++ chains/ethereum/rpc-client/src/client.rs
@@ -6,25 +6,19 @@ use ethers::prelude::*;
 use ethers::providers::JsonRpcClient;
 use ethers::types::U256;
 use futures_timer::Delay;
-use futures_util::future::BoxFuture;
-use futures_util::{FutureExt, Stream};
 use jsonrpsee::core::client::{Subscription, SubscriptionClientT, SubscriptionKind};
 use jsonrpsee::{
     client_transport::ws::WsTransportClientBuilder,
     core::{
-        client::{Client, ClientBuilder, ClientT},
+        client::{Client as JsonRpseeClient, ClientBuilder, ClientT},
         error::Error as JsonRpseeError,
     },
     rpc_params,
 };
-use pin_project::pin_project;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
-use serde_json::value::RawValue;
 use std::fmt::Debug;
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task::{Context, Poll};
 use std::time::Duration;
 use url::Url;
 
@@ -33,13 +27,13 @@ const ETHEREUM_UNSUBSCRIBE_METHOD: &str = "eth_unsubscribe";
 
 // Websocket Client that supports reconnecting
 #[derive(Debug)]
-pub struct JsonRpseeClient {
-    uri: Url,
-    client: ArcSwapOption<Client>,
-    subscriptions: DashMap<U256, Subscription<serde_json::Value>>,
+pub struct EthClient {
+    pub(crate) uri: Url,
+    pub(crate) client: ArcSwapOption<JsonRpseeClient>,
+    pub(crate) subscriptions: DashMap<U256, Subscription<serde_json::Value>>,
 }
 
-impl JsonRpseeClient {
+impl EthClient {
     pub async fn connect(uri: Url) -> Result<Self, Error> {
         let (tx, rx) = WsTransportClientBuilder::default()
             .build(uri.clone())
@@ -52,7 +46,7 @@ impl JsonRpseeClient {
         })
     }
 
-    pub fn client(&self) -> Result<Arc<Client>, Error> {
+    pub fn client(&self) -> Result<Arc<JsonRpseeClient>, Error> {
         let Some(client) = self.client.load().clone() else {
             log::error!("Requested failed, client is reconnecting...");
             return Err(Error::Reconnecting);
@@ -168,7 +162,7 @@ impl JsonRpseeClient {
 }
 
 #[async_trait]
-impl JsonRpcClient for JsonRpseeClient {
+impl JsonRpcClient for EthClient {
     type Error = Error;
 
     async fn request<T, R>(&self, method: &str, params: T) -> Result<R, Self::Error>
@@ -186,123 +180,7 @@ impl JsonRpcClient for JsonRpseeClient {
     }
 }
 
-pub enum SubscriptionStreamState {
-    Subscribed(Subscription<serde_json::Value>),
-    Unsubscribing(BoxFuture<'static, Result<(), JsonRpseeError>>),
-}
-
-#[pin_project]
-pub struct SubscriptionStream {
-    id: U256,
-    failures: u32,
-    state: Option<SubscriptionStreamState>,
-}
-
-impl SubscriptionStream {
-    pub fn new(id: U256, stream: Subscription<serde_json::Value>) -> Self {
-        Self {
-            id,
-            failures: 0,
-            state: Some(SubscriptionStreamState::Subscribed(stream)),
-        }
-    }
-}
-
-impl Stream for SubscriptionStream {
-    type Item = Box<RawValue>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.project();
-        loop {
-            match this.state.take() {
-                Some(SubscriptionStreamState::Subscribed(mut stream)) => {
-                    let result = match stream.poll_next_unpin(cx) {
-                        Poll::Ready(result) => result,
-                        Poll::Pending => {
-                            *this.state = Some(SubscriptionStreamState::Subscribed(stream));
-                            return Poll::Pending;
-                        }
-                    };
-
-                    // Stream is close
-                    let Some(result) = result else {
-                        return Poll::Ready(None);
-                    };
-
-                    // Parse the result
-                    let result = result.and_then(|value| {
-                        serde_json::value::to_raw_value(&value).map_err(JsonRpseeError::ParseError)
-                    });
-
-                    match result {
-                        Ok(value) => {
-                            *this.state = Some(SubscriptionStreamState::Subscribed(stream));
-                            return Poll::Ready(Some(value));
-                        }
-                        Err(error) => {
-                            log::error!(
-                                "Invalid response from subscription error {}: {:?}",
-                                this.failures,
-                                error
-                            );
-                            *this.failures += 1;
-
-                            if *this.failures > 5 {
-                                log::error!("Too many errors, unsubscribing...");
-                                *this.state = Some(SubscriptionStreamState::Unsubscribing(
-                                    stream.unsubscribe().boxed(),
-                                ));
-                            } else {
-                                *this.state = Some(SubscriptionStreamState::Subscribed(stream));
-                            }
-                            continue;
-                        }
-                    }
-                }
-                Some(SubscriptionStreamState::Unsubscribing(mut future)) => {
-                    return match future.poll_unpin(cx) {
-                        Poll::Ready(Ok(_)) => Poll::Ready(None),
-                        Poll::Ready(Err(error)) => {
-                            log::error!("Failed to unsubscribe: {:?}", error);
-                            Poll::Ready(None)
-                        }
-                        Poll::Pending => {
-                            *this.state = Some(SubscriptionStreamState::Unsubscribing(future));
-                            Poll::Pending
-                        }
-                    };
-                }
-                None => {
-                    log::error!("stream must not be polled after being closed`");
-                    return Poll::Ready(None);
-                }
-            }
-        }
-    }
-}
-
-impl PubsubClient for JsonRpseeClient {
-    type NotificationStream = SubscriptionStream;
-
-    /// Add a subscription to this transport
-    fn subscribe<T: Into<U256>>(&self, id: T) -> Result<Self::NotificationStream, Error> {
-        let Some((id, stream)) = self.subscriptions.remove(&id.into()) else {
-            return Err(Error::JsonRpsee {
-                original: JsonRpseeError::InvalidSubscriptionId,
-                message: None,
-            });
-        };
-
-        Ok(SubscriptionStream::new(id, stream))
-    }
-
-    /// Remove a subscription from this transport
-    fn unsubscribe<T: Into<U256>>(&self, _id: T) -> Result<(), Self::Error> {
-        Ok(())
-    }
-}
-
-pub async fn is_connected(client: &Client) -> bool {
+pub async fn is_connected(client: &JsonRpseeClient) -> bool {
     if !client.is_connected() {
         return false;
     }
