+++ chains/ethereum/rpc-client/src/lib.rs
@@ -1,8 +1,10 @@
 mod client;
 mod error;
 mod params;
+mod subscription;
 
-pub use client::{JsonRpseeClient as Client, SubscriptionStream};
+pub use client::EthClient;
+pub use subscription::SubscriptionStream;
 
 pub mod prelude {
     pub use ethers::providers::{JsonRpcClient, PubsubClient, RpcError};
